// src/typings/api/trade.d.ts
export type IMarginCoinItem = {
  id: string
  name: string
  balance: string
  appLogo: string
  autoAddBalance: string
}
// ...
export type MarginCoinArrayType = {
  sort: number
  rate: string
  coinId: number
  coinName: string
  selected: boolean
  currencySymbol: string
}